04-deployment/homework/pyproject.toml
[tool.poetry]
name = "taxi-predictor"
version = "0.1.0"
description = ""
authors = ["ugm2 <unaigaraymaestre@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.8"
scikit-learn = "1.2.2"
pandas = "^2.0.2"
pyarrow = "^12.0.1"
google-cloud-storage = "^2.10.0"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
